Managed to find 81 intersections. Working from main.py.
Filtering happens in mainDetect.py. Very inefficient algorithm at the moment looking for the distance of every intersection to every neighbor and excluding the neighbor if a minimum distance is not given, i.e. it has found more than one point. need to run it three times using a for loop to work properly and it's also complexity n^2. Anyway, the list rows now contains all corners in the format (row, column, (x coord, y coord)). next step is to draw polygons from those with a class-based system. once that is done we just need to be able to run color analysis for each polygon and we^re done <3<3<3<3<3