Major update
Started working on the Square class which takes the chessboard corners and from there makes objects. Still working on that. Cleaned up the rest of the code, moving functioning bits to the right place. Biggest issue is probably still the crappy filter for clustered intersections. Working from Main.py, the functions one and the squareClass one. The corners are now saved in a dictionary where the key is there location in the form [row,column]. I am trying to loop through the dictionary to instantiate all the 64 squares. The current problem is that some don't initialise correctly, as some __init__ variables use cv2 methods that give use information about the centre of the square. These fail sometimes...