Classification of squares
The script is now able to detect squares, order them, and classify whether it's a black or a white square. Next step is to use images of populated chessboards in order to detect which piece has moved and of what color it is. Also need to be able to output a BWE matrix.