Hough lines & contours
Added a filter to hough lines so that it is no longer neccessary to
filter points (this section of the code could be removed in the future).
Sort lines based on how close their y(horizontal) or x(vertical) centres
are.
Contours no longer requires manual calibration: made the assumption that
area/perimeter will always be largest for the chess board